feat: added TTL to recording preview in playlist
