--- frontend/src/scenes/session-recordings/player/utils/ttlUtils.ts
@@ -1,8 +1,0 @@
-import { dayjs } from 'lib/dayjs'
-
-export function calculateTTL(recording_start_time: string, retention_period_days: number): number {
-    const start = dayjs(recording_start_time)
-    const now = dayjs()
-
-    return Math.max(retention_period_days - now.diff(start, 'days'), 0)
-}
