+++ frontend/src/scenes/session-recordings/playlist/SessionRecordingPreview.tsx
@@ -4,11 +4,11 @@ import clsx from 'clsx'
 import { useActions, useValues } from 'kea'
 import { memo } from 'react'
 
-import { IconBug, IconCursorClick, IconKeyboard, IconLive } from '@posthog/icons'
+import { IconBug, IconCursorClick, IconHourglass, IconKeyboard, IconLive } from '@posthog/icons'
 
 import { PropertyIcon } from 'lib/components/PropertyIcon/PropertyIcon'
 import { TZLabel } from 'lib/components/TZLabel'
-import { FEATURE_FLAGS } from 'lib/constants'
+import { FEATURE_FLAGS, SESSION_RECORDINGS_TTL_WARNING_THRESHOLD_DAYS } from 'lib/constants'
 import { LemonCheckbox } from 'lib/lemon-ui/LemonCheckbox'
 import { LemonSkeleton } from 'lib/lemon-ui/LemonSkeleton'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
@@ -24,6 +24,7 @@ import { urls } from 'scenes/urls'
 import { RecordingsQuery } from '~/queries/schema/schema-general'
 import { SessionRecordingType } from '~/types'
 
+import { calculateTTL } from '../utils'
 import { sessionRecordingsListPropertiesLogic } from './sessionRecordingsListPropertiesLogic'
 import {
     DEFAULT_RECORDING_FILTERS_ORDER_BY,
@@ -243,6 +244,12 @@ export const SessionRecordingPreview = memo(
 
         const iconClassNames = 'text-secondary shrink-0'
 
+        const recordingTTL = recording.retention_period_days
+            ? calculateTTL(recording.start_time, recording.retention_period_days)
+            : null
+        const ttlColor =
+            recordingTTL && recordingTTL <= SESSION_RECORDINGS_TTL_WARNING_THRESHOLD_DAYS ? '#f63b3bff' : 'currentColor'
+
         return (
             <DraggableToNotebook href={urls.replaySingle(recording.id)}>
                 <div
@@ -294,6 +301,14 @@ export const SessionRecordingPreview = memo(
                                             <span>{recording.keypress_count}</span>
                                         </span>
                                     </Tooltip>
+                                    {recordingTTL && (
+                                        <Tooltip className="flex items-center" title="Days until recording expires">
+                                            <span className="flex gap-x-0.5">
+                                                <IconHourglass fill={ttlColor} className={iconClassNames} />
+                                                <span style={{ color: ttlColor }}>{recordingTTL}d</span>
+                                            </span>
+                                        </Tooltip>
+                                    )}
                                 </div>
                             </div>
